+++ node/pkg/celo/celoimpl.go
@@ -17,7 +17,7 @@ import (
 
 	celoAbi "github.com/certusone/wormhole/node/pkg/celo/abi"
 	common "github.com/certusone/wormhole/node/pkg/common"
-	ethAbi "github.com/certusone/wormhole/node/pkg/ethereum/abi"
+	ethAbi "github.com/certusone/wormhole/node/pkg/evm/abi"
 
 	"go.uber.org/zap"
 )
