+++ node/pkg/evm/utils_test.go
@@ -1,4 +1,4 @@
-package ethereum
+package evm
 
 import (
 	"testing"
